Improved the footer section
on clicking the address part it will open the location of the college in google map in another tab.
on clicking the phone no. dialer will open you will be able to directly call from your pc/laptop with the linked phone
on clicking the email you will be able to mail directly from your device